* Poshuku: Fixed segfault on delete in empty favorites view.

+++ src/plugins/poshuku/favoritestreeview.cpp
@@ -12,7 +12,8 @@ FavoritesTreeView::~FavoritesTreeView ()
 
 void FavoritesTreeView::keyPressEvent (QKeyEvent *e)
 {
-	if (e->key () == Qt::Key_Delete)
+	if (e->key () == Qt::Key_Delete &&
+			selectionModel ()->currentIndex ().isValid ())
 		emit deleteSelected (selectionModel ()->currentIndex ());
 
 	QTreeView::keyPressEvent (e);
